+++ library/SolverUtils/DriverAdaptive.cpp
@@ -97,7 +97,7 @@ void DriverAdaptive::v_Execute(ostream &out)
     m_equ[0]->DoInitialise();
 
     // Obtain initial time in case a restart was used
-    NekDouble startTime = m_equ[0]->GetFinalTime();
+    NekDouble startTime = m_equ[0]->GetTime();
     m_equ[0]->DoSolve();
 
     // Load session parameters and solver info
@@ -424,7 +424,7 @@ void DriverAdaptive::v_Execute(ostream &out)
         Array<OneD, NekDouble> exactsoln(m_equ[0]->GetTotPoints(), 0.0);
 
         // Evaluate "ExactSolution" function, or zero array
-        m_equ[0]->EvaluateExactSolution(i, exactsoln, m_equ[0]->GetFinalTime());
+        m_equ[0]->EvaluateExactSolution(i, exactsoln, m_equ[0]->GetTime());
 
         NekDouble vL2Error   = m_equ[0]->L2Error(i, exactsoln);
         NekDouble vLinfError = m_equ[0]->LinfError(i, exactsoln);